refactor(assistant): extract rolling container CPU sampler into a shared util (#40862)
